05-OpenAI-For-Exams/07-Streaming/pyproject.toml
[project]
name = "05-weather-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.1",
]
